add ValueTreeState to processor
Note: I started to try to connect the dial and button to the processor using listeners and stuff, but apparently the ValueTreeState method is newer, better, and works better with DAWs?

I've left some stuff in but just commented out in case I want it. I'll go back and delete whatever isn't used after I finish the ValueTreeState